Added: in page notification options (#127)
* Added: in page notification options
Changed: browser sync storage as there is a limit to how manny packages you can send to the cloud (this is done with a timestamp counter) > 10 curently. if threasshold is reached it will delay send and send it one second later. further "adds" are added to the same "batch" and time is reset by 1 sec
Changed so that it wont .set of value already equal to .get

* fixed spelling error.

* implemented sugested changes

* more sugestions fixed